requirements.txt.txt
pyttsx3
SpeechRecognition
pyaudio
pywhatkit
pyjokes
requests
wikipedia
python-docx
python-pptx
customtkinter
psutil
gTTS
